fix(cli): route vt serve through ensureNodeVtDaemonForProject; tear down orphaned graphd on failure (rec 1)
serve previously called the low-level ensureVtDaemonForProject with the wrong arity (deps.resolvePath is not a function), never starting vtd and orphaning graphd. Now routes through the high-level node-ensure entry that builds state/deps, and tears down graphd if vtd-ensure fails. e2e test harness extracted from the >500-line test file.
