+++ packages/systems/voicetree-cli/src/serveOwner.e2e.test.ts
@@ -1,12 +1,13 @@
-// BF-377 — `vt serve` two-ensure wrapper.
+// `vt serve` two-ensure wrapper — black-box e2e spec.
 //
-// Black-box: spawn the real `vt serve` CLI as a child process and assert
-// observable outcomes — stdout ready line, /health owner identity for both
-// daemons, on-disk graphd.owner.json + vtd.owner.json contents, exit code,
-// stderr fragments, cross-process lifetime of the spawned daemons. No
-// internal mocks.
+// Spawns the real `vt serve` CLI as a child process and asserts observable
+// outcomes — stdout ready line, /health owner identity for both daemons,
+// on-disk graphd.owner.json + vtd.owner.json contents, exit code, stderr
+// fragments, cross-process lifetime of the spawned daemons. No internal mocks.
+// The spawn / health / owner-record / teardown effect helpers live in
+// `serveOwner.e2e.harness.ts`.
 //
-// Six scenarios cover the two-ensure wrapper contract:
+// Seven scenarios cover the two-ensure wrapper contract:
 //   1. Cold-start spawns one vt-graphd AND one vt-daemon; both /health probes
 //      prove identity; both owner records on disk.
 //   2. Warm reuse of both pre-existing healthy owners.
@@ -18,284 +19,55 @@
 //      invariant extended to vt-daemon).
 //   6. Idle: a quiet `vt serve` is still alive 2s after ready (guards the
 //      `await new Promise(() => {})` idle path).
+//   7. Induced vt-daemon ensure failure tears down the graph-db daemon this
+//      invocation just launched — no orphan left behind, exit non-zero.
+//
+// The vt-daemon ensure is exercised through the high-level
+// `ensureNodeVtDaemonForProject(runtime, project, caller, options)` entry —
+// both inside `vt serve` (the subject) and in this file's own prewarm setup
+// (`prewarmVtd`). The low-level `ensureVtDaemonForProject(state, deps, …)`
+// takes the single-flight state + deps as its first two positional args and
+// must NOT be called with a `(project, caller, options)` arity.
 
-import {spawn, type ChildProcess} from 'node:child_process'
-import {mkdir, mkdtemp, readFile, rm} from 'node:fs/promises'
-import {createRequire} from 'node:module'
+import {mkdir, mkdtemp, rm} from 'node:fs/promises'
 import {tmpdir} from 'node:os'
-import {dirname, join, resolve} from 'node:path'
-import {fileURLToPath} from 'node:url'
+import {join, resolve} from 'node:path'
 import {afterEach, beforeEach, describe, expect, it} from 'vitest'
-import {
-    ensureGraphDaemonForProject,
-    GraphDbClient,
-    type EnsureGraphDaemonResult,
-} from '@vt/graph-db-client'
-import {
-    ensureVtDaemonForProject,
-    type EnsureVtDaemonResult,
-} from '@vt/vt-daemon-client'
-import {
-    HealthResponseSchema,
-    ownerRecordFile,
-    VtDaemonHealthResponseSchema,
-    type HealthResponse,
-    type OwnerRecord,
-    type VtDaemonHealthResponse,
+import type {EnsureGraphDaemonResult} from '@vt/graph-db-client'
+import type {EnsureVtDaemonResult, VtDaemonClient} from '@vt/vt-daemon-client'
+import type {
+    HealthResponse,
+    OwnerRecord,
+    VtDaemonHealthResponse,
 } from '@vt/graph-db-protocol'
-
-const TEST_FILE_DIR: string = dirname(fileURLToPath(import.meta.url))
-const PACKAGE_DIR: string = resolve(TEST_FILE_DIR, '..')
-const REPO_ROOT: string = resolve(PACKAGE_DIR, '../../..')
-const CLI_ENTRYPOINT: string = join(PACKAGE_DIR, 'src/voicetree-cli.ts')
-// Resolve the tsx CLI through Node's own module lookup. The worktree layout
-// keeps the bulk of dependencies in the main repo's node_modules and only
-// symlinks a tiny .bin subset under per-package node_modules, so a hardcoded
-// node_modules/.bin/tsx path is unreliable here.
-const TSX_REQUIRE = createRequire(import.meta.url)
-const TSX_PACKAGE_DIR: string = dirname(TSX_REQUIRE.resolve('tsx/package.json'))
-const TSX_CLI_PATH: string = join(TSX_PACKAGE_DIR, 'dist', 'cli.mjs')
-const NODE_BIN: string = process.execPath
-const CLI_TSCONFIG: string = join(PACKAGE_DIR, 'tsconfig.json')
-const GRAPHD_SOURCE_BIN: string = join(REPO_ROOT, 'packages/systems/graph-db-server/bin/vt-graphd.ts')
-const VTD_SOURCE_BIN: string = join(REPO_ROOT, 'packages/systems/vt-daemon/bin/vtd.ts')
-// `node <tsx-cli> <script>` is the worktree-portable equivalent of running
-// `tsx <script>` directly — see TSX_CLI_PATH derivation above. Used for both
-// graphd and vtd; the per-daemon ensure clients split on whitespace and
-// append `--project-root <project>` / `--project <project>` respectively.
-const VT_GRAPHD_BIN_OVERRIDE: string = `${NODE_BIN} ${TSX_CLI_PATH} ${GRAPHD_SOURCE_BIN}`
-const VT_DAEMON_BIN_OVERRIDE: string = `${NODE_BIN} ${TSX_CLI_PATH} ${VTD_SOURCE_BIN}`
-
-const SERVE_READY_TIMEOUT_MS: number = 20_000
-const SERVE_EXIT_TIMEOUT_MS: number = 10_000
-const DAEMON_SHUTDOWN_TIMEOUT_MS: number = 5_000
-const SCENARIO_TIMEOUT_MS: number = 60_000
-const ENSURE_TIMEOUT_MS: number = 20_000
-
-type ServeReady = {
-    readonly graphdVerb: 'launched' | 'reused'
-    readonly graphdPort: number
-    readonly graphdPid: number
-    readonly vtdVerb: 'launched' | 'reused'
-    readonly vtdUrl: string
-    readonly vtdPid: number
-}
-
-type ServeHandle = {
-    readonly child: ChildProcess
-    readonly stdoutBuffer: {value: string}
-    readonly stderrBuffer: {value: string}
-    readonly ready: Promise<ServeReady>
-    readonly exited: Promise<{code: number | null; signal: NodeJS.Signals | null}>
-}
-
-function sleep(ms: number): Promise<void> {
-    return new Promise<void>((res) => setTimeout(res, ms))
-}
-
-function buildChildEnv(voicetreeHome: string): Record<string, string> {
-    const merged: Record<string, string | undefined> = {
-        ...process.env,
-        VOICETREE_HOME_PATH: voicetreeHome,
-        TSX_TSCONFIG_PATH: CLI_TSCONFIG,
-        VT_GRAPHD_BIN: VT_GRAPHD_BIN_OVERRIDE,
-        VT_DAEMON_BIN: VT_DAEMON_BIN_OVERRIDE,
-    }
-    delete merged.VT_SESSION
-    delete merged.VOICETREE_TERMINAL_ID
-    const result: Record<string, string> = {}
-    for (const [key, value] of Object.entries(merged)) {
-        if (value !== undefined) {
-            result[key] = value
-        }
-    }
-    return result
-}
-
-function parseServeReady(line: string): ServeReady | null {
-    // Format produced by serve.ts (BF-377):
-    //   "vt serve: graph-db <verb> on http://127.0.0.1:<port> (pid <pid>),
-    //    vt-daemon <verb> on <url> (pid <pid>), project=<path>"
-    const match: RegExpExecArray | null = /^vt serve: graph-db (launched|reused) on http:\/\/127\.0\.0\.1:(\d+) \(pid (\d+)\), vt-daemon (launched|reused) on (\S+) \(pid (\d+)\), project=/m.exec(line)
-    if (!match) return null
-    const graphdVerb: 'launched' | 'reused' = match[1] === 'launched' ? 'launched' : 'reused'
-    const vtdVerb: 'launched' | 'reused' = match[4] === 'launched' ? 'launched' : 'reused'
-    return {
-        graphdVerb,
-        graphdPort: Number(match[2]),
-        graphdPid: Number(match[3]),
-        vtdVerb,
-        vtdUrl: match[5],
-        vtdPid: Number(match[6]),
-    }
-}
-
-function spawnServe(args: string[], voicetreeHome: string): ServeHandle {
-    const child: ChildProcess = spawn(
-        NODE_BIN,
-        [TSX_CLI_PATH, CLI_ENTRYPOINT, 'serve', ...args],
-        {
-            cwd: REPO_ROOT,
-            env: buildChildEnv(voicetreeHome),
-            stdio: ['ignore', 'pipe', 'pipe'],
-        },
-    )
-    const stdoutBuffer = {value: ''}
-    const stderrBuffer = {value: ''}
-
-    let readyResolve: (value: ServeReady) => void = () => {}
-    let readyReject: (reason: Error) => void = () => {}
-    const ready: Promise<ServeReady> = new Promise<ServeReady>((res, rej) => {
-        readyResolve = res
-        readyReject = rej
-    })
-
-    let exitedResolve: (value: {code: number | null; signal: NodeJS.Signals | null}) => void = () => {}
-    const exited: Promise<{code: number | null; signal: NodeJS.Signals | null}> = new Promise((res) => {
-        exitedResolve = res
-    })
-
-    let readyResolved: boolean = false
-    const readyTimer: NodeJS.Timeout = setTimeout(() => {
-        if (readyResolved) return
-        readyReject(new Error(
-            `vt serve did not become ready within ${SERVE_READY_TIMEOUT_MS}ms.\n`
-            + `stdout:\n${stdoutBuffer.value}\nstderr:\n${stderrBuffer.value}`,
-        ))
-    }, SERVE_READY_TIMEOUT_MS)
-
-    child.stdout?.on('data', (chunk: Buffer): void => {
-        stdoutBuffer.value += chunk.toString('utf8')
-        if (!readyResolved) {
-            const parsed: ServeReady | null = parseServeReady(stdoutBuffer.value)
-            if (parsed !== null) {
-                readyResolved = true
-                clearTimeout(readyTimer)
-                readyResolve(parsed)
-            }
-        }
-    })
-    child.stderr?.on('data', (chunk: Buffer): void => {
-        stderrBuffer.value += chunk.toString('utf8')
-    })
-
-    child.on('exit', (code: number | null, signal: NodeJS.Signals | null): void => {
-        clearTimeout(readyTimer)
-        if (!readyResolved) {
-            readyReject(new Error(
-                `vt serve exited before ready (code=${code} signal=${signal})\n`
-                + `stdout:\n${stdoutBuffer.value}\nstderr:\n${stderrBuffer.value}`,
-            ))
-        }
-        exitedResolve({code, signal})
-    })
-
-    return {child, stdoutBuffer, stderrBuffer, ready, exited}
-}
-
-async function stopServe(handle: ServeHandle): Promise<{code: number | null; signal: NodeJS.Signals | null}> {
-    if (handle.child.exitCode === null && handle.child.signalCode === null) {
-        handle.child.kill('SIGTERM')
-    }
-    const killTimer: NodeJS.Timeout = setTimeout(() => {
-        if (handle.child.exitCode === null && handle.child.signalCode === null) {
-            handle.child.kill('SIGKILL')
-        }
-    }, SERVE_EXIT_TIMEOUT_MS)
-    try {
-        return await handle.exited
-    } finally {
-        clearTimeout(killTimer)
-    }
-}
-
-async function readOwnerRecord(project: string, daemonKind: 'graphd' | 'vtd'): Promise<OwnerRecord | null> {
-    try {
-        const raw: string = await readFile(ownerRecordFile.pathFor(project, daemonKind), 'utf8')
-        return ownerRecordFile.decode(raw)
-    } catch (err) {
-        if ((err as NodeJS.ErrnoException).code === 'ENOENT') return null
-        throw err
-    }
-}
-
-async function fetchGraphdHealth(port: number): Promise<HealthResponse> {
-    const response: Response = await fetch(`http://127.0.0.1:${port}/health`)
-    expect(response.ok, `unexpected graphd /health status ${response.status}`).toBe(true)
-    const body: unknown = await response.json()
-    const parsed = HealthResponseSchema.safeParse(body)
-    expect(parsed.success, `graphd /health body did not parse: ${JSON.stringify(body)}`).toBe(true)
-    if (!parsed.success) throw new Error('unreachable')
-    return parsed.data
-}
-
-async function fetchVtdHealth(url: string): Promise<VtDaemonHealthResponse> {
-    const response: Response = await fetch(`${url}/health`)
-    expect(response.ok, `unexpected vtd /health status ${response.status}`).toBe(true)
-    const body: unknown = await response.json()
-    const parsed = VtDaemonHealthResponseSchema.safeParse(body)
-    expect(parsed.success, `vtd /health body did not parse: ${JSON.stringify(body)}`).toBe(true)
-    if (!parsed.success) throw new Error('unreachable')
-    return parsed.data
-}
-
-function isPidAlive(pid: number): boolean {
-    try {
-        process.kill(pid, 0)
-        return true
-    } catch (err) {
-        return (err as NodeJS.ErrnoException).code === 'EPERM'
-    }
-}
-
-async function shutdownGraphd(port: number): Promise<void> {
-    const client: GraphDbClient = new GraphDbClient({baseUrl: `http://127.0.0.1:${port}`})
-    await client.shutdown().catch(() => undefined)
-}
-
-async function waitForDaemonExit(pid: number, timeoutMs: number): Promise<boolean> {
-    const deadline: number = Date.now() + timeoutMs
-    while (Date.now() < deadline) {
-        if (!isPidAlive(pid)) return true
-        await sleep(50)
-    }
-    return false
-}
-
-async function ensureCleanProject(project: string): Promise<void> {
-    const graphdRecord: OwnerRecord | null = await readOwnerRecord(project, 'graphd')
-    if (graphdRecord !== null && graphdRecord.port !== null) {
-        await shutdownGraphd(graphdRecord.port)
-        if (graphdRecord.pid) {
-            await waitForDaemonExit(graphdRecord.pid, DAEMON_SHUTDOWN_TIMEOUT_MS)
-        }
-    }
-    const vtdRecord: OwnerRecord | null = await readOwnerRecord(project, 'vtd')
-    if (vtdRecord !== null && vtdRecord.pid) {
-        // vtd has no GraphDbClient analogue; signal-terminate the pid.
-        try {
-            process.kill(vtdRecord.pid, 'SIGTERM')
-        } catch {
-            // already gone
-        }
-        await waitForDaemonExit(vtdRecord.pid, DAEMON_SHUTDOWN_TIMEOUT_MS)
-    }
-    for (const file of [
-        'graphd.owner.json',
-        'graphd.lock',
-        'graphd.port',
-        'vtd.owner.json',
-        'vtd.spawn.lock',
-        'vtd.cooldown.json',
-        'auth-token',
-        'rpc.port',
-    ]) {
-        await rm(join(project, '.voicetree', file), {force: true}).catch(() => undefined)
-    }
-}
+import {
+    DAEMON_SHUTDOWN_TIMEOUT_MS,
+    ENSURE_TIMEOUT_MS,
+    SCENARIO_TIMEOUT_MS,
+    VT_DAEMON_BIN_FAILING,
+    VT_DAEMON_BIN_OVERRIDE,
+    VT_GRAPHD_BIN_OVERRIDE,
+    ensureCleanProject,
+    fetchGraphdHealth,
+    fetchVtdHealth,
+    graphdHealthReachable,
+    isPidAlive,
+    prewarmGraphd,
+    prewarmVtd,
+    readOwnerRecord,
+    shutdownGraphd,
+    sleep,
+    spawnServe,
+    stopServe,
+    waitForDaemonExit,
+    waitForExclusiveRefusal,
+    waitForVtdFailureExit,
+    type ServeHandle,
+    type ServeReady,
+} from './serveOwner.e2e.harness'
 
 describe.skipIf(process.env.CI_SANDBOX === '1')(
-    'BF-377 vt serve two-ensure wrapper',
+    'vt serve two-ensure wrapper',
     () => {
         let root: string
         let project: string
@@ -368,13 +140,13 @@ describe.skipIf(process.env.CI_SANDBOX === '1')(
                 process.env.VT_GRAPHD_BIN = VT_GRAPHD_BIN_OVERRIDE
                 process.env.VT_DAEMON_BIN = VT_DAEMON_BIN_OVERRIDE
 
-                const graphdPrewarm: EnsureGraphDaemonResult = await ensureGraphDaemonForProject(
+                const graphdPrewarm: EnsureGraphDaemonResult = await prewarmGraphd(
                     project,
                     'test',
                     {bin: VT_GRAPHD_BIN_OVERRIDE, timeoutMs: ENSURE_TIMEOUT_MS},
                 )
                 expect(graphdPrewarm.launched).toBe(true)
-                const vtdPrewarm: EnsureVtDaemonResult = await ensureVtDaemonForProject(
+                const vtdPrewarm: EnsureVtDaemonResult<VtDaemonClient> = await prewarmVtd(
                     project,
                     'test',
                     {bin: VT_DAEMON_BIN_OVERRIDE, timeoutMs: ENSURE_TIMEOUT_MS},
@@ -414,7 +186,7 @@ describe.skipIf(process.env.CI_SANDBOX === '1')(
                 // exclusive check passes, allowing the vt-daemon-named
                 // exclusive-conflict error to surface (graph-db's check
                 // runs first by construction in serve.ts).
-                const vtdPrewarm: EnsureVtDaemonResult = await ensureVtDaemonForProject(
+                const vtdPrewarm: EnsureVtDaemonResult<VtDaemonClient> = await prewarmVtd(
                     project,
                     'test',
                     {bin: VT_DAEMON_BIN_OVERRIDE, timeoutMs: ENSURE_TIMEOUT_MS},
@@ -451,7 +223,7 @@ describe.skipIf(process.env.CI_SANDBOX === '1')(
                 process.env.VT_GRAPHD_BIN = VT_GRAPHD_BIN_OVERRIDE
                 process.env.VT_DAEMON_BIN = VT_DAEMON_BIN_OVERRIDE
 
-                const graphdPrewarm: EnsureGraphDaemonResult = await ensureGraphDaemonForProject(
+                const graphdPrewarm: EnsureGraphDaemonResult = await prewarmGraphd(
                     project,
                     'test',
                     {bin: VT_GRAPHD_BIN_OVERRIDE, timeoutMs: ENSURE_TIMEOUT_MS},
@@ -515,32 +287,41 @@ describe.skipIf(process.env.CI_SANDBOX === '1')(
             },
             SCENARIO_TIMEOUT_MS,
         )
+
+        it(
+            'induced vt-daemon failure: tears down the just-launched graph-db daemon (no orphan), exits non-zero',
+            async () => {
+                // Point vt serve at a vt-daemon bin that exits non-zero
+                // immediately and never publishes an owner, with a short
+                // ensure timeout so the vt-daemon ensure fails deterministically.
+                // graph-db ensure runs FIRST and succeeds, launching a fresh
+                // graphd — the orphan candidate. The fix must tear it down.
+                const handle: ServeHandle = spawnServe(
+                    ['--project', project],
+                    voicetreeHome,
+                    VT_DAEMON_BIN_FAILING,
+                )
+                serveHandles.push(handle)
+
+                const exitResult = await waitForVtdFailureExit(handle)
+                expect(exitResult.code, 'vt serve must exit non-zero on vt-daemon ensure failure').not.toBe(0)
+                expect(handle.stderrBuffer.value).toMatch(/failed to ensure vt-daemon owner/)
+
+                // The graph-db daemon this invocation launched must have been
+                // torn down — assert on the observable side effect (its
+                // /health is no longer reachable and its pid is gone), not on
+                // any internal call.
+                const graphdRecord: OwnerRecord | null = await readOwnerRecord(project, 'graphd')
+                if (graphdRecord !== null && graphdRecord.port !== null) {
+                    const stillReachable: boolean = await graphdHealthReachable(graphdRecord.port)
+                    expect(stillReachable, 'graph-db daemon must NOT be left orphaned after vt-daemon failure').toBe(false)
+                    if (graphdRecord.pid) {
+                        const exited: boolean = await waitForDaemonExit(graphdRecord.pid, DAEMON_SHUTDOWN_TIMEOUT_MS)
+                        expect(exited, 'orphaned graph-db pid must have exited').toBe(true)
+                    }
+                }
+            },
+            SCENARIO_TIMEOUT_MS,
+        )
     },
 )
-
-async function waitForExclusiveRefusal(
-    handle: ServeHandle,
-): Promise<{code: number | null; signal: NodeJS.Signals | null}> {
-    return Promise.race([
-        handle.exited,
-        handle.ready.then(
-            (): never => {
-                throw new Error(
-                    `--exclusive unexpectedly became ready:\n${handle.stdoutBuffer.value}`,
-                )
-            },
-            (): never => {
-                throw new Error(
-                    `--exclusive ready promise rejected:\n${handle.stderrBuffer.value}`,
-                )
-            },
-        ),
-    ]).catch(async (err: Error) => {
-        if (err.message.startsWith('--exclusive ready promise rejected')) {
-            // Ready promise rejects when child exits before ready — exactly
-            // the path --exclusive takes against a pre-existing owner.
-            return await handle.exited
-        }
-        throw err
-    })
-}
